requirements.txt
flask==3.0.3
python-dotenv==1.0.1
supabase==2.5.3
qrcode[pil]==7.4.2
Pillow==10.3.0
reportlab==4.2.2
Werkzeug==3.0.3
gunicorn==22.0.0